fix(test): stop skipping Ollama SSRF regressions in the default install
test_ssrf_guarded_url_components.py carries a module-level
pytest.importorskip("lfx_bundles") so the moved provider shims do not fail at
collection now that lfx-bundles is opt-in (#13869). That guard is correct for
deepseek/glean/homeassistant/huggingface/litellm/lmstudio/xai, but it also
covered the two Ollama tests -- and Ollama graduated out of the metapackage
into lfx-ollama, which IS a default langflow dependency. Both tests therefore
skipped in an install that has the components present and working.

Split them into test_ssrf_guarded_ollama_components.py, guarded on lfx_ollama.
They cannot fold into test_chatollama_component.py or
test_ollama_embeddings_component.py: both of those classes carry an autouse
disable_ssrf_protection fixture, which is the very protection under test here.
The new file replicates the original's autouse enable_ssrf_protection fixture
so it does not depend on ambient env.

Patches are anchored via sys.modules[Component.__module__] rather than a fixed
dotted string, matching the convention in test_chatollama_component.py -- these
modules are reachable under the lfx.components.ollama.* shim, the canonical
lfx_ollama.components.ollama.*, and the runtime _lfx_ext.* ext-loader copy.

Verified in the default (no lfx_bundles) install: 2 passed, and 101 passed
alongside the sibling Ollama suites.
